Set default img at 1000px if original width is greater
This avoids the need for the original image to be copied in the
case when it is only needed as the default (when srcset isn't
supported), particularly when it is large. Instead we reuse the
1000px image which will be generated for the srcset.

If the image is smalled than 1000px do nothing and the original
will also be part of the sourceset at its original width.